DOM functionality consolidated and typed supported
